refactor: rename iimage.h to image.h to match its type
The header contained class Image (a value type), not an interface, so
the i prefix was misleading and would silently break consumers if
renamed after 1.0. Updated 13 in-tree includers; itarget.h is left
alone since it actually contains class ITarget.
