+++ include/moth_graphics/graphics/asset_context.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "moth_graphics/graphics/ifont.h"
-#include "moth_graphics/graphics/iimage.h"
+#include "moth_graphics/graphics/image.h"
 #include "moth_graphics/graphics/itexture.h"
 #include "moth_graphics/utils/rect.h"
