+++ tests/src/test_spritesheet.cpp
@@ -1,5 +1,5 @@
 #include "moth_graphics/graphics/spritesheet.h"
-#include "moth_graphics/graphics/iimage.h"
+#include "moth_graphics/graphics/image.h"
 #include "moth_graphics/graphics/itexture.h"
 #include "moth_graphics/utils/rect.h"
 #include "moth_graphics/utils/vector.h"
